+++ src/main/java/br/com/sis/rh/apiprogramaformacao/core/repository/RemuneracaoInstrutorRepository.java
@@ -1,5 +1,6 @@
 package br.com.sis.rh.apiprogramaformacao.core.repository;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -18,5 +19,5 @@ public interface RemuneracaoInstrutorRepository extends JpaRepository<Remuneraca
 	@Query(value = "select sum(vlr_hora * qtd_hora) from TB_REMUNERACAO_INSTRUTOR ri join TB_PROGRAMA p "
 		+ "ON ri.codigo_instrutor_fk = p.cpf_Instrutor where nome = ?1 AND nome_turma = ?2 "
 		+ "AND data_lancamento >= ?3 AND data_lancamento <= ?4", nativeQuery = true)
-	Double calcularSalarioInstrutoresPeriodo(String nomePrograma, String nomeTurma, LocalDate dataInicio, LocalDate dataFim);
+	BigDecimal calcularSalarioInstrutoresPeriodo(String nomePrograma, String nomeTurma, LocalDate dataInicio, LocalDate dataFim);
 }
