+++ src/main/java/br/com/sis/rh/apiprogramaformacao/core/util/FormatadorDeNumeroDecimalUtil.java
@@ -1,5 +1,6 @@
 package br.com.sis.rh.apiprogramaformacao.core.util;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 
 import org.springframework.stereotype.Service;
@@ -9,7 +10,7 @@ public class FormatadorDeNumeroDecimalUtil {
 
 	private DecimalFormat formatter = new DecimalFormat("#,##.00");
 	
-	public String formatarDecimal(Double nota) {
+	public String formatarDecimal(BigDecimal nota) {
 		String notaFormatada = formatter.format(nota);
 		return notaFormatada;
 	}
